Pull binding control into BindingHandler class +PluginIgnoreAttribute for ignoring plugin types
