+++ TabletDriverPlugin/Attributes/PluginIgnoreAttribute.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace TabletDriverPlugin.Attributes
+{
+    [AttributeUsage(AttributeTargets.Class | AttributeTargets.Struct | AttributeTargets.Interface)]
+    public class PluginIgnoreAttribute : Attribute
+    {
+    }
+}
